+++ cpp/src/scheduler/action/Action.h
@@ -5,26 +5,29 @@
  ******************************************************************************/
 #pragma once
 
-#include "Event.h"
+#include "../resource/Resource.h"
 
 
 namespace zilliz {
 namespace milvus {
 namespace engine {
 
-class StartUpEvent : public Event {
+class Action {
 public:
-    explicit
-    StartUpEvent(std::weak_ptr<Resource> resource)
-        : Event(EventType::START_UP, std::move(resource)) {}
+    /*
+     * Push task to neighbour;
+     */
+    static void
+    PushTaskToNeighbour(const ResourceWPtr &self);
 
-    inline std::string
-    Dump() const override {
-        return "<StartUpEvent>";
-    }
 
-    friend std::ostream &operator<<(std::ostream &out, const StartUpEvent &event);
+    /*
+     * Pull task From neighbour;
+     */
+    static void
+    PullTaskFromNeighbour(const ResourceWPtr &self);
 };
+
 
 }
 }