+++ cpp/src/scheduler/event/CopyCompletedEvent.h
@@ -18,6 +18,14 @@ public:
     CopyCompletedEvent(std::weak_ptr<Resource> resource, TaskTableItemPtr task_table_item)
         : Event(EventType::COPY_COMPLETED, std::move(resource)),
           task_table_item_(std::move(task_table_item)) {}
+
+    inline std::string
+    Dump() const override {
+        return "<CopyCompletedEvent>";
+    }
+
+    friend std::ostream &operator<<(std::ostream &out, const CopyCompletedEvent &event);
+
 public:
     TaskTableItemPtr task_table_item_;
 };