+++ cpp/unittest/scheduler/normal_test.cpp
@@ -1,13 +1,14 @@
 #include "scheduler/ResourceFactory.h"
 #include "scheduler/ResourceMgr.h"
 #include "scheduler/Scheduler.h"
+#include "scheduler/task/TestTask.h"
+#include "utils/Log.h"
 #include <gtest/gtest.h>
 
 
 using namespace zilliz::milvus::engine;
 
-TEST(normal_test, DISABLED_test1) {
-
+TEST(normal_test, test1) {
     // ResourceMgr only compose resources, provide unified event
     auto res_mgr = std::make_shared<ResourceMgr>();
     auto disk = res_mgr->Add(ResourceFactory::Create("disk", "ssd"));
@@ -23,17 +24,35 @@ TEST(normal_test, DISABLED_test1) {
 
     res_mgr->Start();
 
-    auto task1 = std::make_shared<XSearchTask>();
-    auto task2 = std::make_shared<XSearchTask>();
-    if (auto observe = disk.lock()) {
-        observe->task_table().Put(task1);
-        observe->task_table().Put(task2);
-        observe->task_table().Put(task1);
-        observe->task_table().Put(task1);
-    }
-
     auto scheduler = new Scheduler(res_mgr);
     scheduler->Start();
 
-    while (true) sleep(1);
+    auto task1 = std::make_shared<TestTask>();
+    auto task2 = std::make_shared<TestTask>();
+    auto task3 = std::make_shared<TestTask>();
+    auto task4 = std::make_shared<TestTask>();
+    if (auto observe = disk.lock()) {
+        observe->task_table().Put(task1);
+        observe->task_table().Put(task2);
+        observe->task_table().Put(task3);
+        observe->task_table().Put(task4);
+        std::cout << "disk:" << std::endl;
+        std::cout << observe->task_table().Dump() << std::endl;
+    }
+
+    sleep(5);
+
+    if (auto observe = disk.lock()) {
+        std::cout << "disk:" << std::endl;
+        std::cout << observe->task_table().Dump() << std::endl;
+    }
+    if (auto observe = cpu.lock()) {
+        std::cout << "cpu:" << std::endl;
+        std::cout << observe->task_table().Dump() << std::endl;
+    }
+    scheduler->Stop();
+    res_mgr->Stop();
+
+    ASSERT_EQ(task1->load_count_, 1);
+    ASSERT_EQ(task1->exec_count_, 1);
 }